Remove `AlphaDBMysql class`
